+++ src/card/cards/combat_constraints.rs
@@ -5,3 +5,4 @@ use crate::CardDefinitionId;
 pub const ELDER_LAND_WURM: CardDefinitionId = CardDefinitionId(1795);
 pub const DWARVEN_SOLDIER: CardDefinitionId = CardDefinitionId(1796);
 pub const BATTERING_RAM: CardDefinitionId = CardDefinitionId(1797);
+pub const SINGING_TREE: CardDefinitionId = CardDefinitionId(1798);